Add files via upload
The Z-score, also known as the standard score, tells us the deviation of a data point from the mean by expressing it in terms of standard deviations above or below the mean.
The formula for calculating a z-score is: z = (x – μ)/ σ
Where:
x: is the test value
μ: is the mean
σ: is the standard value

What are some the application of Z-Score
-Z-scores are often used for feature scaling to bring different features to a common scale. Normalizing features ensures that they have zero mean and unit variance, which can be beneficial for certain machine learning algorithms, especially those that rely on distance measures.
-Z-scores can be used to identify outliers in a dataset. Data points with Z-scores beyond a certain threshold (usually 3 standard deviations from the mean) may be considered outliers.
-Z-scores can be used in anomaly detection algorithms to identify instances that deviate significantly from the expected behaviour.
-Z-scores can be applied to transform skewed distributions into more normal distributions.
When working with regression models, Z-scores of residuals can be analyzed to check for homoscedasticity (constant variance of residuals).
Z-scores can be used in feature scaling by looking at their standard deviations from the mean.


Z-Scores have various applications, such as in finance for portfolio analysis, education for standardized testing, health for clinical assessments, and more. They provide a standardized measure for comparing and interpreting data.